Value substitution, initial commit.
Implemented value substition.

Created a boolean test to roughly compare two XML documents. Using this
to verify the output of a Stencil template.

Made the resolver a separate package that is supplied to the Stencil
package. The Stencil package will define a browser based resolver if
none is provided. There may end up being a compile step, so that if a
developer can use AMD or some other alternative resolver, they don't
have to carry the weight of the default resolver.

Closes #5.
Closes #2.
Closes #1.