Implemented `require` attribute.
This commit raises a concern that a require attribute is going to
complicate the template engine needlessly. We could do with a require
element and then all of the Stencil directives would be specified by
element.

Elements are easier to process. Stencil element operations operate on
the whole element, including attributes. We detect a Stencil element,
processes it, and then resume with its next sibling, or parent if it is
the last child. If the processing is asynchronous, the callback can
resume with the next node in document order.

When some attributes are special, we need to linger over elements and
their attributes, potentially processing them asynchronously. If we
process an attribute asynchronously, we need to resume with the same
element, but the next attribute. Not as simple as our element only
traversal.

Also, attributes are nodes, but they are not linked, so we need to
iterate over them using an index. When I set out to write this commit,
I'd misremembered attributes in the DOM, thinking them to be linked as
siblings, hoping to obtain some reuse from the traversal logic.

In this commit I've moved to a closure style. All the traversal code and
operator implementation is wrapped in a closure. The functions are no
longer passing around a `descent` object.

I've implemented a function cache, so that when we encounter a selector
to evaluate, we use an existing selector if we can find one that shares
the same code and the same variables in scope. Don't get clever and
think that you can rely on just the source as a key.

This commit uses version `0.1.11` of `xmldom`. I submitted a patch to
address [attribute namespaces
issues](https://github.com/jindw/xmldom/commit/dc0ecd772673dc4d1d0c746154002620bd01c8f4).
Nice to be giving `xmldom` some exercise. There does not appear to an
alternative W3C DOM to `xmldom` for use with Node.js.

The `resolver` function is external to the Stencil template engine. I've
created one specifically for Node.js. I'll consider how to provision one
for the browser later on.