Tempaltes working, plus lots of cruft.
Got templates to work. Able to switch contexts from a calling template
to a caller template, creating a whole new scope, then returning to
caller's scope when invoking the caller's blocks. Had problems with the
stack, detecting that it was empty multiple times and triggering the
callback of the included template multiple times.

`xmlify` now accepts a stack. Most state is kept in the stack at the top
element. Using `shift` and `unshift` for `push` and `pop` so that the
top element is always convienently at `0`.

Moved resolution of resources into a `fetch` function that could be used
to resolve templates requested by the caller as well as by include
statements.

Changed `require` to use the `xmlns` prefix name as the variable name
for assignent and created a `require` protocol as discussed in 17. TK

Updated Proof to 0.0.10.

Instead of simply returning `false`, I updated `t/compare.js` to call an
`abend` function that simply returns false. This allows me to jump into
that `abend` function and throw an error when I need to debug why XML is
not comparing when I can't find the differences visually. Was having
problems with `xmldom` where it was getting confused about attributes
because the `localName`s of the attributes were `null`, so the documents
didn't compare, even though they looked find when serialized.

Added `jindw/xmldom` as a dependency through `git`. I'm now a committer
on the project, so I'm able to push changes into the `jindw/xmldom`
repository, instead of having to wait for them to get pushed to NPM.

I'd considered using my own mirror of `jindw/xmldom`, but that would
mean I'd have to keep it around for a while, so that older versions of
Stencil would build. This obviates the needs for difficult decision to
maintain a repo, or wait on each publish.

Added a lot of additional cruft. Got tired of sorting it during commits
and I'm pretty sure I'm going to use it all eventually.