Implement `each`.
Changed the `xmlify` function to visit the children of the node at the
base of the stack, rather than the node itself. This means passing in
the document node in instead of the document element node when
traversing the root document or an included document. That is the only
real change. When we visit blocks, visting the root of the block is a
no-op in our current set of tests.

Closes #7.