Normalize relative URLs.
Created a URL relativizer and normalizer. Moved resolution out of
reslover. Resolver is now really just a loader.

The URL resolution is simple and not intended to work with every valid
URL, just the reasonable subset of valid URLs necessary to get the
project off the ground. Support will be added as the complaints roll in.

The URL resolver attempts to account for query strings and bookmarks. No

Not going to implement a search path for includes. So, #12 is invalid.

Closes #22.
Closes #13.
Closes #4.
Closes #3.