Implement function cache by source.
Cache by source. Any invocation of an expression had better have the
stuff it needs in scope, or else there is going to be an error. Maybe
not the first time, but down the road for sure. This is an unpleasantry
we're going to suffer upon the application developer for the sake of the
minification, simplicity.

Function caching is now compiler function wide, by source.

    File                             Raw         Min         Zip     Min/Zip
    index.js                 12674/12.38  5003/04.89  3871/03.78  2157/02.11

Closes #9.