Implement HTML serializer.
In the browser, we're going to import nodes using `Document.adoptNode`.
On the server side we do need to serialize, but it only needs to work on
the server. We can have a minimalist serialzer.

Because we remove CDATA sections in our generator, we can rely on a
regular expression to fixup void elements, and close empty elements. Our
serializer will only work for XML that describes HTML5 and has no CDATA
sections.

Added some notes to the `README.md`.

    File                             Raw         Min         Zip     Min/Zip
    index.js                 12436/12.14  4777/04.67  3724/03.64  2030/01.98

Closes #10.