Push JSON update into DOM.
Now able to update the DOM based on changes to the underlying JSON
document used to generate the DOM from a Stencil.

No API for this yet. Simply wanted to see the mechanics of the push.

The generator now returns an object. As there is no API, the object just
contains the raw state and a method named `update`. All this object
really needs to expose is a destructor, to dispose of the listening
snippets when it is unloaded or replaced.

There is probably some way to have garbage collection. Each call to the
generator builds one document, a live document. If a snippet is
refreshed, but it's node that is no longer a member of the document, no
action is taken and the snippet is removed from the set of snippets.