Flatten evaluate, dynamic function.
Sorting out the workings of a pull-to-push function. Within the
`index.js` the beginnings of pull-to-push have already been implemented.
We'd extended the `require` directive to load a function that acted as a
callback when the module was a JSON provider.

That block of code was updated to look for the name `dynamic` attached
to the callback.

Flattened out evaluate. The special `callback` function passed to the
generated expression function has been removed; expressions will never
be asynchronous. Asynchronous comes when the expression returns a
function.