DOM based reference tracing.
Instead of tracing which JSON values are referenced using the proxy
based on property getter functions, we're running the templates a second
time and tracking what values are referenced. We can then do a compare
on the values to see if we insert the generated snippet or discard it.

Note that we're not comparing the generated DOM against the existing
DOM. That would create false positives when we're inserting HTML,
because the HTML parsing step might produce different output on the
server than in the browser. Plus, it would hurt to parse a big glob of
HTML only to discard it. Plus, comparing HTML generated DOM is slower
than a string compare of the source.

We are going to implement deferred parsing of HTML since it is obviously
going to be a performance hit. Any HTML source values that need to be
converted to DOM will be gathered into a structure. Only if the region
is dirty is the HTML source parsed and inserted into the DOM.

Nested regions are an unaddressed challenge. Serialization of templates
is now broken.

What's working is the push into the registered callback.