Pull to push for single each.
Implemented pull to push for a single template each. Evaluations are now
separated by a level of indirection. The stack indexes into an array of
evaluations. Keeps us from cloning JSON trees when we capture the stack.

In the stack, we no longer capture the context, but the series of
evaluations. We rebuild the context for each evaluation. This may be
slow, but for each directives, we can cache a calculated context.

Had a realization too, that we don't actually have to navigate the dome
to find the sources for our evaluation. We can cache them with the
values. The only thing we lose now from our proxied JSON tree is the
ability to know if an each directive touches a particular JSON tree.

Stencil is a template language, not a query language. We're not often
going to be performing joins of JSON trees within a template. It will
mostly be the case that an each directive will iterate over a particular
JSON tree. The dependency mechanism I was building with the JSON tree
proxies is of dubious value. For the most part, we'd have to walk the
entire tree anyway. Hard to imagine, at this point, if there was an
opportunity for performance improvements, and under what circumstances.
I don't feel like I'm losing anything other than complexity.