Include tags without descending into template.
It was my intent to allow  tag library authors to include resources with
`xmlns` directives on the `include` element.

Descending into the document and walking it using our existing template
logic causes these requires and includes to be evaluated. I thought the
implementation would simply grab the short stack created by this
descent, then start a descent into a tag primed with the `include`
element stack. However, the `include` element might require a dynamic
data source, which would generate a push end point. That one push end
point would push into any document that included the template.

One push end point updating multiple documents is depressing.

Instead, we gather up the tags. When we want to dive into a tag, we
start from the include element, skip all tags except for the one we want
to visit, and have the include and require directives play out as if
they were a part of the document.

The next step is to implement dive from `include`. With this commit I've
extracted the available tag names into a collection attached to the
template data structure.