Sort out use of `document` and `nodes`.
Removed the variable name `doc` where it was found. Instead I'm naming
my documents `document` and the contents of a document or the results of
a template `nodes`.

I removed `document` as a keyword from the default syntax highlighting
for JavaScript in Vim. This was why I was avoiding the variable name
`document`. The variable `doc` looks silly and does not reflect the
stature of that which it references.