Created descent data structure.
Wrapped the details of a descent into a document into a data structure.
Reduces the signature of our workhorse `xmlify` function.

In the stack in the `descent` structure, `evaluations` has been renamed
`assignments`. In the `stencil` structure, `evaluations` has been
renamed `dynamics`. The `createDescent` method helps to make the
creation of a descent less fiddly, more declarative.

Will still manipulate the `stack` directly. The `descent` gathers our
descent state, it does not encapsulate it.

Renamed the `wrap` function to `createElement`.

Brought in the `validator` function from Strata.

Closes #28.
Closes #27.