Tag library pull, evaluated attributes.
Implemented the tag library for a pull, but have not yet implemented
push. Descent into tags works both for imported tags and tags declared
by tag blocks.

Moved the construction of the `included` directive wrapped `tag`
declaration into the library `include` function. Tags are now ready to
go when they are encountered. The same logic that descents through the
`include` directive can be used to descent through a `tag` directive
that is a block child.

Implemented `parameter` which allows a library to inject its context
into the context of the caller when rendering a caller's block.

Implemented evaluated attributes. I'd included them in my tag library
example stencils, having forgotten that I'd forgotten to implement them.

Closes #53.
Closes #32.