Implement tag libraries.
Now able to import a simple tag library and have it participate in
pull-to-push dynamic regions. Our serialized stacks now include
references to any tags defined by include or block entry.

The descent object has been converted into a linked list of descents.
When we generate a context for an evaluation, we go up the list of
descents including all stacks for any descent matching the URL of the
current descent.

We're doing a better job of tracking dirty sections, inheriting
properties marked as dirty when we travel through a block or tag.

We've created two collections of assignments now, one for static
entries, one for dynamic entries. Our serialized descents will now only
contain a reference to a data structure.

We're assigning ids to elements in templates that define operations,
both user defined tags and Stencil directive elements. The id is a
combination of template id and document order. XML DOM has been patched
to fix a bug where it did not include the `documentElement` when
`Document.getElementsByTagName` was invoked; the `package.json`
references the release with this update.

    File                             Raw         Min         Zip     Min/Zip
    index.js                 34152/33.35 12484/12.19  8557/08.36  4491/04.39

Closes #38.