Rebuild instance structure from DOM comments.
We're now able to rebuild the instance structure for values and
attributes using data stored in comment nodes in the DOM. Currently, the
comment markers store only the id of the directive that inserted the DOM
nodes. A JSON tree is exported at the end of the document that includes
the count of nodes and characters inserted. Future commits will explore
storing the instance meta information in the comment markers directly.