Add template depth to path.
Correctly generate the path by adding the template depth. This allows
templates to become reentrant when we support included layouts and tag
libraries.

I changed the logic for nesting directive parts in directive paths, but
none of the tests generate nested directive paths yet, so we're not
going to see any changes. The way it was written, a value directive that
was a child of an if directive would be nested in the if directive, so
the path for the child directive would include the id of the if
directive as a parent. This isn't necessary. The value directive is
unique in the template, the if directive does nothing to distinguish it.

When we introduce an each directive, we are going to add the user
specified id to the directive part. This allows us to reuse the
directive parts for any branches, values or attribute markers within the
each block which will repeat.

If the user doesn't not provide a id for a series, we can simply use the
index of the array. I'm moving away from the notion of having Stencil
itself perform some sort of check to see if an update is needed, a
dependency system. It will be enough to push data into the DOM and have
the DOM rewrite itself.

Closes #57.