No longer storing instance data in array.
Instance data will be stored in a tree. Each member of a collection will
have a key based on an id derived from the collection object or else the
index of the object in the collection. The array was an early notion of
how to represent multiple instances of a directive.