Store instances in tree.
Now storing instance data -- element counts, character counts and
markers -- in a tree. This is so that when we implement conditionals we
can free instance data by unlinking an instance from an tree and letting
it fall out of scope. All of the instance data of the child directives
will be only be referenced through the instance data of their parent
directives, so the unlinked instance data is garbage collected all of
the instance data of descendant directives will also be garbage
collected.