Implemented `if`.
Implemented an `if` with a test for when if is true.

Compilation now unfolds directives, moving thier contents out into the
document as next siblings, and then replacing the directive with a
marker. This makes the document consistent for the `rewrite` function.

The handlers for the directives now accept the current descent object so
that an if directive can clear the copy of the child directives when is
the condition is false so that the child directives are not evaluated.