Implement false `if` directive.
Correctly reconstituting a Stencil document that contains an `if`
directive.

Implemented `if` false.

Noticing now that I do want to nest directives underneath an `if`
statement in a directive path, even though each directive underneath an
`if` statement is unique. The new path key is an array that navigates a
tree of instance data, so long keys don't bother me as much, since each
part of the path exists only once in the tree.

Yes, long hash keys bothered me, even though as long as they may get,
they'd never reach a maximum key length. I thought it would be ugly if
it ever got displayed or used in debugging.

Now the instance data is a tree, so we could free instance data with the
directives underneath an `if` directive, if the instance tree was
parallel to the directive tree. The shortened paths do not create a
parallel instance tree.