Make `comments` accept array `path`.
The `comments` function used the path as a means to communicate the
range of a directive body. The path was specified as an array of plain
objects, the top most object in the path containing the elements and
characters of the directive body. This made it difficult to reuse the
`comments` function in other parts of the code that only had an array of
strings.

Now the caller receives the range of a directive by passing in a plain
object as the first parameter. It is populated by `comments` function
when the given node is a directive marker comment. This outbound
function parameter is only ever used by the `comments` function itself
to implement an order traversal of the DOM. Other callers will pass in
an anonymous plain object, ignoring the result.