Implement `each`.
Implemented `each`. Not certain how to handle an update that would
reorder the elements. I will probably go through the array checking to
see if the element exists in the current document. If it exists, then I
need to test if it is linked into the document at the correct location,
relinking it if it is not in the correct location.

The `reconstitute` function is choking on the document.

The `scavenge` function needs to be updated to strip out any each object
ids from the path. Once an `if` or `each` direction is nested in an
`each`, the path to the instance in the generated page will be different
from the path to the directive prototype in the scrap page.

See #60.