Single pass convert, templates as fragments.
Now converting a template into a prototype page in a single pass. The
body of the template is moved into a fragment so that the root of the
template can contain multiple elements, which is necessary to compile
tag libraries where the `module` tag and the `tag` tags are flattened
into directive comments, all children of the same parent node.

Lost the ability to assert that assert that an if` directive is
correctly structured. May reintroduce later. May create a template
linter instead. Maybe leave it as is if it turns out to be a goof that
is easy to spot visually.

File                             Raw         Min         Zip     Min/Zip
redux.js                 25027/24.44  8491/08.29  6690/06.53  3151/03.08