Use object equality to test node identity.
I was using a special comment injected into the DOM prior to rewriting a
directive body to mark the end of a directive body. Directive body is
then calculated by searching for the special comment. The special
comment was thought to be necessary because of a memory of `isSameNode`.
The `isSameNode` function has been mercifully put to rest.

The program to an interface dogma leads developers to believe that they
have to extend their interface to include any possible implementation;
the dogmatic approach leads one to believe that the interface cannot
identify the object, so stark is the "separation of concerns."