Swap items set.
To reduce the surprise of the side effect of `comments` adding the item
id to the item id set of the prototype I now assign the `items` property
of the prototype to the `items` local variable for use as a difference
set, then create a new empty set in the `items` property or the
prototype where the new set of items is built. When `comments` adds an
id to the item set of the prototype it now merely does what the `each`
directive handler has already done.

Closes #73.