Implement out-of-band element.
Simplify the directive body marker logic by implementing a start and an
end marker that sandwiches a directive creating an out-of-band element
that does not effect CSS or jQuery selectors, that can account for DOM
rewriting by jQuery libraries.

The `vivify` function now takes note of both the beginning comment and
the ending comment, so that DOM manipulation is implemented as iteration
along siblings, looking for the end element using object comparison.
Developers can add new elements to the directive body and the directive
body can still be identified by the start and end markers.

I am not going to implement `recount` so this closes #74 as invalid.

I've been able to serialize this minimal implementation of tags for a
while now and I don't seen any issues with tag serialization going
forward, so this closes #54 also.

File                             Raw         Min         Zip     Min/Zip
redux.js                 23931/23.37  7592/07.41  6395/06.25  2988/02.92